Create new test file for test_get_multiple_2000_store_messages
